refactor(widgets/post): extract PostCard, PostMeta, PostActions from PostList
