+++ src/app/App.tsx
@@ -6,7 +6,7 @@ import { NotFound } from '@/pages/NotFound.tsx';
 import { Profile } from '@/pages/Profile.tsx';
 import { AuthLayout } from '@/layouts/AuthLayout.tsx';
 import { MainLayout } from '@/layouts/MainLayout.tsx';
-import { PostList } from '@/widgets/PostList/ui/PostList.tsx';
+import { PostList } from '@/features/PostList/ui/PostList.tsx';
 import { Route, Routes } from 'react-router-dom';
 import { Auth } from '@/pages/Auth.tsx';
