+++ src/features/PostList/ui/PostList.styled.ts
@@ -1,0 +1,24 @@
+import styled from 'styled-components';
+export const PostContainer = styled.div`
+  width: 57rem;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+  justify-content: start;
+  gap: 40px 0;
+`;
+
+export const SysMessage = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 24px;
+  font-weight: bold;
+  width: 100%;
+  height: 100%;
+  padding: 40px 0;
+`;
+
+
+
