+++ src/widgets/PostForm/ui/PostFormActions.tsx
@@ -39,7 +39,7 @@ export const PostFormActions: React.FC<Props> = ({ onClose }) => {
   return (
     <Container>
       <Cancel type="button" onClick={onClose}>
-        Cancel
+        Отменить
       </Cancel>
       <Submit type="submit">Submit</Submit>
     </Container>
