+++ src/widgets/PostForm/ui/PostFormFields.tsx
@@ -5,8 +5,8 @@ import { RiAiGenerateText } from 'react-icons/ri';
 import { GoImage } from 'react-icons/go';
 import { useInsertEmoji } from '@/shared/lib/hooks/useInsertEmoji.ts';
 import { EmojiSelector, InlineActionToolbar, Input } from '@/shared/ui';
-import { PostFormPartialProps } from '@/widgets/PostForm/lib/types.ts';
 import { ValidationError } from '@/shared/ui/ValidationError.styled.ts';
+import { PostFormPartialProps } from '@/widgets/PostForm/lib/types.ts';
 
 const Container = styled.div`
   position: relative;
